feat(assets): add /v1/assets/ranks endpoint
Per-version rank metadata (tier, localized name, badge image URLs, color)
built from the patch's merged localization map. Reuses the shared
localization loader, Language enum, and version resolver. Snapshot-tested
against the existing citadel_main_{english,russian} KV1 fixtures.

Also fixes ObjectiveRegen.out_of_combat_health_regen to f64 — the source
field is m_flOutOfCombatHealthRegen and current data ships fractional values.

+++ api/src/services/assets/versions/npc_units.rs
@@ -241,7 +241,7 @@ struct RawIntrinsicModifiers {
 #[derive(Debug, Deserialize, Clone, Copy)]
 struct RawObjectiveRegen {
     #[serde(default, rename = "m_flOutOfCombatHealthRegen")]
-    out_of_combat_health_regen: Option<i64>,
+    out_of_combat_health_regen: Option<f64>,
     #[serde(default, rename = "m_flOutOfCombatRegenDelay")]
     out_of_combat_regen_delay: Option<f64>,
 }
@@ -610,7 +610,7 @@ pub(crate) struct IntrinsicModifiers {
 #[derive(Debug, Serialize, Clone, ToSchema)]
 pub(crate) struct ObjectiveRegen {
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub out_of_combat_health_regen: Option<i64>,
+    pub out_of_combat_health_regen: Option<f64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub out_of_combat_regen_delay: Option<f64>,
 }
